brand-analyzer-app/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
databricks-sdk==0.38.0
databricks-sql-connector==3.6.0
